Refactor previous changes, fix deleting account bug
+ Everything is now in authentication_state.dart
+ Removed notifications; directly call the function via a class instance.
