+++ app/lib/view/account_switcher/account_switcher.dart
@@ -3,7 +3,7 @@ import 'package:sph_plan/core/database/account_database/account_db.dart';
 import 'package:sph_plan/view/login/auth.dart';
 
 import '../../core/sph/sph.dart';
-import '../../models/startup.dart';
+import '../../utils/authentication_state.dart';
 import 'account_list_tile.dart';
 
 class AccountSwitcher extends StatefulWidget {
@@ -41,7 +41,7 @@ class _AccountSwitcherState extends State<AccountSwitcher> {
                   }
                   await sph!.session.deAuthenticate();
                   await accountDatabase.setNextLogin(account.id);
-                  reset(context);
+                  authenticationState.reset(context);
                 }, dbID: account.id,
               );
             },
